requirements.txt
requests
python-dotenv
console-menu
time